Upload of final project thesis.
Abstract -

The purpose of this study is to identify if software can be used as aid for ageing at home. Ireland has what is called an ageing population, in 2016 the amount of people aged 65+ increased by 19.1% compared to 2011 census [12]. Coupled with the ageing population, Ireland’s infrastructure currently needs an investment of 2bn [7] to accommodate the older persons. This study aims to identify if it possible to live in your own home as opposed to nursing homes when you are aged 65+ using software as an aid.