Code base for end of year project.
This folder contains majority code of my final year project, thesis is also view-able.

This code is not runnable as I have not uploaded the full project as this is something I am continuing to develop. 
This code is placed here to make it possible for others to view my work.